Optimise check, add better unit tests
Check only the beginning of content instead of all.
Remove warc unit test because the issue we are trying to resolve appears
only in arc files.
Add arc unit test with DELETED content (should skip it).
Add arc unit test with DELETED not in the beginning of content (should create
CDX).